Fixed #1959 - Added column protection for environment schema migration

+++ lib/puppet/rails/database/003_add_environment_to_host.rb
@@ -1,9 +1,13 @@
 class AddEnvironmentToHost < ActiveRecord::Migration
     def self.up
-        add_column :hosts, :environment, :string
+        unless ActiveRecord::Base.connection.columns(:hosts).collect {|c| c.name}.include?("environment")
+            add_column :hosts, :environment, :string
+        end
     end
     
     def self.down
-        remove_column :hosts, :environment
+        if ActiveRecord::Base.connection.columns(:hosts).collect {|c| c.name}.include?("environment")
+            remove_column :hosts, :environment
+        end
     end
 end
